chore: Extract macros to doctest/parts/macros.h
